pyproject.toml
[project]
name = "neuro-forge"
version = "0.0.1"
readme = "README.md"
requires-python = ">=3.9"
license = {text = "CeCILL-B"}
classifiers = [
    "Intended Audience :: Developers",
    "License :: OSI Approved",
    "Development Status :: 5 - Production/Stable",
    "Environment :: Console",
    "Operating System :: OS Independent",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Natural Language :: English",
]

dependencies = [
    "fire", 
    "click", 
    "rich", 
    "toml", 
    "pyaml", 
    "gitpython",
]


[project.urls]
repository = "https://github.com/neurospin/neuro-forge"

[build-system]
requires = ["setuptools >= 61.0"]
build-backend = "setuptools.build_meta"

[project.scripts]
neuro-forge = "neuro_forge:main"
soma-forge = "neuro_forge.soma_forge.commands:cli"

[tool.setuptools]
include-package-data = true

[tool.setuptools.packages.find]
exclude = ["recipes"]

[tool.pixi.project]
channels = ["conda-forge", "bioconda"]
platforms = ["linux-64"]

[tool.pixi.dependencies]
fire = "*"
click ="*"
rich = "*"
toml = "*"
pyaml = "*"
gitpython = "*"
conda-build = "*"
rattler-build = "*"
git = "*"
pip = "*"
